redux action 코드를 slice.js로 이동
